pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]

[project]
name = "dolfinx_adjoint"
version = "0.2.0"
description = "Automatic differentation compatible with DOLFINx"
authors = [{ name = "Jørgen S. Dokken", email = "dokken@simula.no" }]
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
      "fenics-dolfinx>=0.10.0",
      "pyadjoint-ad>=2025.10.0",
      "typing_extensions; python_version < '3.11'",
]


[project.optional-dependencies]
test = ["pytest"]
dev = ["pdbpp", "ipython", "mypy", "ruff"]
docs = [
      "jupyter-book",
      "jupytext",
      "moola@git+https://github.com/funsim/moola.git",
      "pandas",
      "pyvista[all]",
]
all = ["dolfinx_adjoint[test]", "dolfinx_adjoint[dev]", "dolfinx_adjoint[docs]"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
# Folders to exclude
exclude = ["docs/", "build/"]
# Folder to check with mypy
files = ["src", "tests"]

[tool.ruff]
src = ["src", "docs", "tests", "examples"]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
      # Pyflakes
      "F",
      # Pycodestyle
      "E",
      "W",
      # isort
      "I001",
]


[tool.ruff.lint.isort]
known-first-party = ["dolfinx_adjoint"]
known-third-party = [
      "basix",
      "dolfinx",
      "ffcx",
      "ufl",
      "gmsh",
      "numpy",
      "pytest",
      "pyadjoint",
]
section-order = [
      "future",
      "standard-library",
      "mpi",
      "third-party",
      "first-party",
      "local-folder",
]

[tool.ruff.lint.isort.sections]
"mpi" = ["mpi4py", "petsc4py"]
